+++ test/src/test.rs
@@ -1,8 +1,8 @@
 mod arg_matcher_test;
-mod pickable_test;
 mod basic_test;
 mod multi_arg_test;
 mod multi_value_test;
+mod pickable_test;
 mod pos_matcher_test;
 mod priority_test;
 mod route_test;
@@ -10,3 +10,4 @@ mod style_test;
 mod value_extra_test;
 mod value_flag_test;
 mod value_string_test;
+mod value_valueless_test;